DBZ-6646 - Added integration test to test offset storage in MySQL containers.
